Provide meaningful names to just ui tests.
