--- client/test/Foo.js
@@ -1,22 +1,0 @@
-import React, { PropTypes } from 'react';
-
-const propTypes = {};
-
-const defaultProps = {};
-
-class Foo extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return (
-      <div className="foo">Similar Listings Service</div>
-    );
-  }
-}
-
-Foo.propTypes = propTypes;
-Foo.defaultProps = defaultProps;
-
-export default Foo;
